add update cli comnnad and remove model cli command
