+++ src/dev/lab/lab.ts
@@ -1,20 +1,6 @@
-import { orm, Respository } from '../../orm'
-import { Customer, Order } from '../../model'
+import { orm } from '../../orm'
+import { CustomerRespository, Customer, Order, Categories, Customers } from '../../schemas/northwind'
 
-class CustomerRespository extends Respository<Customer, QryCustomer> {
-	constructor (database: string) {
-		super('Customers', database)
-	}
-
-	async findByName (firstName: string, lastName: string):Promise<Customer[]> {
-		return await this.query().filter(p => p.name === firstName).execute({ firstName: firstName, lastName: lastName })
-	}
-
-	// eslint-disable-next-line @typescript-eslint/ban-types
-	public lambda (func: Function):string {
-		return orm.lambda(func).complete('')
-	}
-}
 // class OrderRepository extends Respository<Order, QryOrder> {
 // constructor (database: string) {
 // super('Orders', database)
@@ -29,14 +15,21 @@ class CustomerRespository extends Respository<Customer, QryCustomer> {
 		customer.name = 'a'
 		customer.orders.push(new Order())
 		const name = 'a'
+
+		await customerRepository.insert().execute(customer)
+
 		let complete = customerRepository.insert(() => ({ name: name })).include(p => p.orders).complete()
 		console.log(complete)
 		const query = (name: string, description: string) => Categories.insert(() => ({ name: name, description: description }))
 		complete = orm.lambda(query).complete('northwind')
 		console.log(complete)
 
-		const result = await orm.lambda(query).execute({ name: 'test1', description: 'test1' })
-		console.log(result)
+		const query2 = () => Customers
+		complete = orm.lambda(query2).complete('northwind')
+		console.log(complete)
+
+		// const result = await orm.lambda(query).execute({ name: 'test1', description: 'test1' })
+		// console.log(result)
 		// complete = customerRepository.update(p => [p.name, p.id]).include(p => p.orders).complete()
 		// console.log(complete)
