+++ src/orm/cli.ts
@@ -7,7 +7,8 @@ import { SyncCommand } from './cli/sync'
 import { ImportCommand } from './cli/import'
 import { ExportCommand } from './cli/export'
 import { RunCommand } from './cli/run'
-import { ModelCommand } from './cli/model'
+// import { ModelCommand } from './cli/model'
+import { UpdateCommand } from './cli/update'
 
 // eslint-disable-next-line no-unused-expressions
 yargs
@@ -19,7 +20,8 @@ yargs
 	.command(new ImportCommand())
 	.command(new ExportCommand())
 	.command(new RunCommand())
-	.command(new ModelCommand())
+	// .command(new ModelCommand())
+	.command(new UpdateCommand())
 	.recommendCommands()
 	.demandCommand(1)
 	.strict()
