+++ src/orm/cli/init.ts
@@ -44,49 +44,13 @@ export class InitCommand implements CommandModule {
 			if (config.app.configFile === undefined) {
 				config.app.configFile = 'lambdaorm.yaml'
 			}
-			const db = orm.lib.completeConfig(config, database, dialect, connection)
+			orm.lib.completeConfig(config, database, dialect, connection)
+			// write lambdaorm config
 			await orm.lib.writeConfig(config)
-
-			// create initial structure
-			await Helper.createIfNotExists(path.join(config.app.workspace, config.app.src))
-			await Helper.createIfNotExists(path.join(config.app.workspace, config.app.data))
-			await Helper.copyFile(path.join(__dirname, './../sintaxis.d.ts'), path.join(config.app.workspace, config.app.src, 'sintaxis.d.ts'))
-
-			// if the package.json does not exist create it
-			const packagePath = path.join(config.app.workspace, 'package.json')
-			if (!await Helper.existsPath(packagePath)) {
-				await Helper.writeFile(packagePath, JSON.stringify({ dependencies: {} }, null, 2))
-			}
-
-			// if there is no tsconfig.json create it
-			const tsconfigPath = path.join(config.app.workspace, 'tsconfig.json')
-			if (!await Helper.existsPath(tsconfigPath)) {
-				const tsconfigContent = orm.lib.getTypescriptContent()
-				await Helper.writeFile(tsconfigPath, JSON.stringify(tsconfigContent, null, 2))
-			}
-
-			// install typescript if not installed.
-			const typescriptLib = await orm.lib.getLocalPackage('typescript', config.app.workspace)
-			if (typescriptLib === '') {
-				await Helper.exec('npm install typescript -D', config.app.workspace)
-			}
-
-			// install lambdaorm if it is not installed.
-			const lambdaormLib = await orm.lib.getLocalPackage('lambdaorm', config.app.workspace)
-			if (lambdaormLib === '') {
-				await Helper.exec('npm install lambdaorm', config.app.workspace)
-			}
-			// if the library is not installed locally corresponding to the dialect it will be installed
-			const lib = orm.lib.getLib(db.dialect)
-			const localLib = await orm.lib.getLocalPackage(lib, config.app.workspace)
-			if (localLib === '') {
-				await Helper.exec(`npm install ${lib}`, config.app.workspace)
-			}
-			// if the library is not installed locally corresponding to the dialect it will be installed
-			const globalLib = await orm.lib.getGlobalPackage(lib)
-			if (globalLib === '') {
-				await Helper.exec(`npm install ${globalLib} -g`, config.app.workspace)
-			}
+			// create structure
+			await orm.lib.createStructure(config)
+			// add libraries for dialect
+			await orm.lib.addDialects(config)
 		} catch (error) {
 			console.error(`error: ${error}`)
 		}
