+++ src/schemas/northwind/customerRespository.ts
@@ -1,0 +1,9 @@
+import { Respository } from '../../orm'
+import { QryCustomer } from './model'
+
+export class CustomerRespository extends Respository<QryCustomer> {
+	constructor (database: string) {
+		super('Customers', database)
+	}
+	// Add your methods here
+}
